requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
reportlab==4.1.0
customtkinter==5.2.2
Pillow==10.3.0
pyinstaller==6.6.0
deep-translator==1.11.4
